added delete functionality

+++ controllers/documentController.ts
@@ -3,7 +3,11 @@ import { STATUS_CODES } from "../constants/constants";
 import { AppDataSource } from "../db";
 import { MulterRequest, RequestQuery } from "../interfaces";
 import { Document } from "../models";
-import { getS3SignedURL, uploadToAWS } from "../helpers/S3Utils";
+import {
+  deleteDocumentFromAWS,
+  getS3SignedURL,
+  uploadToAWS,
+} from "../helpers/S3Utils";
 import { getRandomString } from "../helpers/helperUtils";
 
 const createDocument = async (req: Request, res: Response) => {
@@ -57,11 +61,23 @@ const getAllDocuments = async (
 const deleteDocument = async (req: Request, res: Response) => {
   const documentRepository = AppDataSource.getRepository(Document);
   const { id } = req.params;
-  try {
-    await documentRepository.delete(Number(id));
-    res.status(STATUS_CODES.success).send("document deleted Successfully");
-  } catch (error) {
-    console.log(error);
+
+  const result:Document | null  = await documentRepository.findOneBy({ id: Number(id) });
+  let lesson: Document;
+
+  if (result !== null) {
+    lesson = result;
+    // use lesson here
+    await deleteDocumentFromAWS(lesson.documentURL);
+    try {
+      await documentRepository.delete(Number(id));
+      res.status(STATUS_CODES.success).send("document deleted Successfully");
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    // handle the null case
+    console.log("Null result");
   }
 };
